requirements.txt
fastapi
uvicorn
python-telegram-bot
requests